# server/swagger.yaml
openapi: 3.0.3
info:
  title: Rocko Backend API Docs
  version: 1.0.1
servers:
  - url: https://backend.dev.rocko.cloud
  - url: http://localhost:5000
  - url: https://backend.prod.rocko.cloud
...
paths:
...
  /addUser:
    post:
        tags:
          - user
        summary: ""
        description: ""
        requestBody:
         content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddUser'
            application/xml:
              schema:
                $ref: '#/components/schemas/AddUser'
            application/x-www-form-urlencoded:
              schema:
                $ref: '#/components/schemas/AddUser'
        responses:
          '200':
            description: Data successfully saved in users table!
          '400':
            description: Invalid status value
          '403':
            description: Cannot create user
  /updateUser:
    post:
      tags:
        - user
      summary: ""
      description: ""
      requestBody:
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                    example: aceruling@gmail.com
                  phone:
                    type: string
                    example: 333-343-434
      responses:
          '200':
            description: Data successfully saved in users table!
          '400':
            description: Invalid value
          '403':
            description: Cannot create user
      security:
        - bearerAuth: [] 
  /updateCountry:
    patch:
      tags:
        - user
      summary: ""
      description: ""
      requestBody:
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                    example: example@mail.com
                  country:
                    type: string
                    example: United States
                  ip:
                    type: string
                    example: 98.97.87.21
      responses:
          '200':
            description: Country info updated successfully saved.
          '400':
            description: Invalid status value
          '403':
            description: Cannot create user
      security:
        - bearerAuth: [] 
  /users:
    post:
      tags:
        - user
      summary: ""
      description: ""
      requestBody:
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                    example: example@mail.com
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'          
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '400':
          description: Invalid status value
        '401':
          description: Unauthorized Invalid email
      security:
        - bearerAuth: [] 
  /userid:
    post:
      tags:
        - user
      summary: ""
      description: ""
      requestBody:
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                    example: example@mail.com
      responses:
        '200':
          description: successful operation
          content:
            application/json:
               schema:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 1
        '400':
          description: Invalid status value
        '401':
          description: Unauthorized Invalid email
      security:
        - bearerAuth: [] 
...
components:
  schemas:
...
    AddUser:
      type: object
      properties:
          email:
            type: string
            example: example@mail.com
          wallet_address:
            type: number
            example: 0x8dBD033bE555b6081e5bCf8662677fF23E118c9B
          country_origin:
            type: string
            example: United States
          ipaddress_lastlogin:
            type: string
            example: Below
          active:
            type: boolean
            example: 
    User:
      type: object
      properties:
          id:
            type: integer
            example: 1
          email:
            type: string
            example: example@mail.com
          phone: 
            type: string
            example: 3333343434
          phone_verified:
            type: boolean
            example: 0
          wallet_address:
            type: number
            example: 0x8dBD033bE555b6081e5bCf8662677fF23E118c9B
          inactive:
            type: boolean
            example: 0
          readonly:
            type: boolean
            example: 0
          mfa_enabled: 
            type: string
          country_origin:
            type: string
            example: United States
          country_lastlogin:
            type: string
            example: United States
          ipaddress_lastlogin:
            type: string
            example: Below
          is_vpn:
            type: boolean
            example: 0
          create_time:
            type: string
            example: 2023-12-19T05:37:59.000Z
          modified_time:
            type: string
            example: 2023-12-19T05:37:59.000Z
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT 